feature: get expense by (#3)
* Add GET endpoint for expense retrieval by ID in OpenAPI

This update introduces a GET endpoint under `/expenses/{expenseId}` to allow retrieval of an expense by its ID. The endpoint includes security via BearerAuth and specifies responses for success (200), not found (404), and unauthorized (401). Additionally, a new property `ddd` of type string has been added to the `Expense` schema.

* Add utility methods for error and result creation

Introduce a static method in Error to create an error with a given message. Also, add a utility method in Result to create a failure result from a single error instance. This enhances code readability and reduces redundancy.

* Rename misnamed test class for CreateExpenseUseCase

Corrected the filename and class name from "CreateExpenseUseCasaeTest" to "CreateExpenseUseCaseTest". This ensures consistency and clarity in test naming, reflecting the intended use case accurately.

* Implement findById method in QueryExpenseUseCaseImpl

Added a new FindByIdExpenseUseCase interface and implemented the findById method in QueryExpenseUseCaseImpl. This method retrieves an expense by its UUID and returns a Result object, encapsulating either the found expense or an error message if not found.

* Add support for fetching expense by ID

Introduced a new endpoint `getExpense` in `ExpensesApiController` to fetch an expense by its ID. Added `FindByIdExpenseUseCase` and `ExpenseDtoMapper` to handle the transformation and business logic.

* Add test for finding expense by ID

Implemented a test case to verify fetching an expense by its ID using REST API. This includes assertions to ensure the response status is OK and the returned expense matches the expected ID.

* Remove duplicate import statements

This change eliminates redundant import declarations for `OffsetDateTime` and `Optional` in the `QueryExpenseUseCaseImpl` class. This improves code cleanliness and reduces potential confusion for future maintenance.